requirements.txt
# === Runtime ===
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-filter>=24.2,<25.0
djangorestframework-simplejwt>=5.3,<6.0
djoser
drf-spectacular>=0.27,<0.30
drf-spectacular-sidecar

pytest>=8.0,<9.0
pytest-django>=4.8,<5.0

ruff>=0.6,<1.0
black>=24.8.0
pre-commit>=3.7,<4.0

# Type hints
django-stubs>=5.0,<6.0
djangorestframework-stubs>=3.15,<4.0

# Prod server (consider in future)
# gunicorn>=22.0,<23.0
